refactor(cart): make cart look a bit better
