+++ packages/frontend/src/features/cart/components/cart-products-note.tsx
@@ -33,12 +33,7 @@ export const CartProductsNote = () => {
 
   return (
     <>
-      <Button
-        onClick={() => setOpen(true)}
-        size="sm"
-        variant="outline"
-        className="text-primary/70"
-      >
+      <Button onClick={() => setOpen(true)} variant="outline" className="text-primary/70">
         <Edit />
       </Button>
       <Modal
