+++ packages/frontend/src/features/cart/components/cart-root.tsx
@@ -1,6 +1,12 @@
 "use client";
 
-import { Sheet, SheetContent, SheetTitle, SheetTrigger } from "@/src/components/ui/sheet";
+import {
+  Sheet,
+  SheetContent,
+  SheetFooter,
+  SheetTitle,
+  SheetTrigger,
+} from "@/src/components/ui/sheet";
 import { Button } from "@/src/components/ui/button";
 import { ShoppingCart } from "lucide-react";
 
@@ -8,10 +14,12 @@ import { useEffect, useState } from "react";
 
 import { useAppDispatch } from "@/src/store";
 
-import { CartContent } from "@/src/features/cart/components/cart-content";
-import { PopoverTitle } from "@/src/features/cart/components/cart-title";
+import { CartContent } from "@/src/features/cart/ui/cart-content";
 
 import { cartMethods } from "@/src/features/cart/store/cart-slice";
+import { Typography } from "@/src/components/ui/typography";
+
+import { CartFooter } from "@/src/features/cart/components/cart-footer";
 
 export const CartTrigger = () => {
   const dispatch = useAppDispatch();
@@ -32,12 +40,16 @@ export const CartTrigger = () => {
           Кошик
         </Button>
       </SheetTrigger>
-      <SheetContent withoutClose className="flex flex-col">
-        <SheetTitle>
-          {/* <SheetClose /> */}
-          <PopoverTitle />
+      <SheetContent className="flex flex-col">
+        <SheetTitle className="pr-4 -mt-3.5">
+          <Typography variant="h2" className="text-lg font-semibold">
+            Кошик
+          </Typography>
         </SheetTitle>
         <CartContent />
+        <SheetFooter>
+          <CartFooter />
+        </SheetFooter>
       </SheetContent>
     </Sheet>
   );
