+++ packages/frontend/src/features/cart/ui/cart-content.tsx
@@ -1,9 +1,6 @@
 import { ScrollArea } from "@/src/components/ui/scroll-area";
-import { CartClearButton } from "@/src/features/cart/components/cart-clear-button";
-import { CartEmpty } from "@/src/features/cart/components/cart-empty";
+import { CartEmpty } from "@/src/features/cart/ui/cart-empty";
 import { CartProducts } from "@/src/features/cart/components/cart-items";
-
-import { CartOrderButton } from "@/src/features/cart/components/cart-order-button";
 
 import { useAppSelector } from "@/src/store";
 
@@ -17,10 +14,6 @@ export const CartContent = () => {
           <ScrollArea className="flex-1 mb-4">
             <CartProducts items={items} />
           </ScrollArea>
-          <section className="flex items-center gap-2">
-            <CartClearButton />
-            <CartOrderButton className="flex-1" />
-          </section>
         </>
       ) : (
         <CartEmpty />
